requirements.txt

google-adk==1.0.0
mcp==1.9.1
deprecated

google-generativeai==0.8.5
python-dotenv==1.1.0
